Extract account switch button into a component
Extracted most of the account switch button into its own component.
There is still some left over code in the parent component.
Specifically the submitConfigure function is currently passed into
the new AccountSwitch function. Ideally, I would like to see this
imported straight into the AccountSwitch sub-component. Binding
the action has proved difficult.
